feat (operator_tracker): no content / created status code + lib
